[tests] Move CanDriver to modm-test/mock as it implements modm::Can
